backend/requirements/base.txt
# Core Django
Django==4.2.23
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Redis & Celery
redis==5.0.1
celery==5.3.6
django-celery-beat==2.5.0
django-redis==5.4.0

# Authentication
PyJWT==2.8.0
djangorestframework-simplejwt==5.3.1
firebase-admin==6.3.0
django-allauth==0.59.0

# API Documentation
drf-spectacular==0.27.0

# Utilities
python-decouple==3.8
python-dateutil==2.8.2
pytz==2023.3

# Validation
django-phonenumber-field[phonenumbers]==7.3.0

# File handling
Pillow==10.2.0
django-storages[google]==1.14.2

# Payment integrations
stripe==8.5.0
mercadopago==2.2.1

# Communication
twilio==8.11.0
requests>=2.25.0,<3.0.0
resend>=0.7.0

# CFDI Integration
# facturapi==0.2.0  # Commented out due to dependency conflict

# IP Geolocation
ipinfo==4.4.3

# Monitoring
sentry-sdk==1.39.1
django-health-check==3.17.0
psutil==5.9.8

# Security & Rate Limiting
django-ratelimit==4.1.0
bleach==6.1.0

# Analytics
mixpanel==4.10.0

# Data processing
pandas==2.1.4
scikit-learn==1.3.2
numpy==1.26.2

# Security
cryptography==41.0.7
django-recaptcha==4.0.0

# ASGI Server
gunicorn==21.2.0
uvicorn[standard]==0.25.0

# Django Extensions
django-extensions==3.2.3
django-filter==23.5

# Logging
python-json-logger==2.0.7

# CLI Utilities
colorama==0.4.6